# parser_json: Improve performance by reducing function call